# Improving performance of EndpointSlice controller metrics cache