#### What type of PR is this?
/kind bug
/kind cleanup

#### What this PR does / why we need it:
This fixes a performance issue where the EndpointSlice controller metrics cache could consume a significant proportion of overall kube-controller-manager CPU utilization by performing inefficient operations within a lock. This becomes especially noticeable when there are a large number of Services in a cluster (many thousands) and the resync period results in all Services being requeued. The increased network programming latency during this event can result in downtime if any Services change while the controller workqueue is backed up.

#### Special notes for your reviewer:
Comparing benchmark output from master and current PR:

```
→ benchstat old.txt new.txt
name                             old time/op  new time/op  delta
UpdateServicePortCache100-12      276µs ± 0%     1µs ± 0%   ~     (p=1.000 n=1+1)
UpdateServicePortCache1000-12     328µs ± 0%     1µs ± 0%   ~     (p=1.000 n=1+1)
UpdateServicePortCache10000-12    547µs ± 0%     1µs ± 0%   ~     (p=1.000 n=1+1)
UpdateServicePortCache100000-12  5.01ms ± 0%  0.00ms ± 0%   ~     (p=1.000 n=1+1)
```

```
→ benchcmp old.txt new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                    old ns/op     new ns/op     delta
BenchmarkUpdateServicePortCache100-12        275804        698           -99.75%
BenchmarkUpdateServicePortCache1000-12       328298        720           -99.78%
BenchmarkUpdateServicePortCache10000-12      546786        726           -99.87%
BenchmarkUpdateServicePortCache100000-12     5005505       881           -99.98%
```

#### Does this PR introduce a user-facing change?
```release-note
An inefficient lock in EndpointSlice controller metrics cache has been reworked. Network programming latency may be significantly reduced in certain scenarios, especially in clusters with a large number of Services.
```

/cc @mborsz @freehan @aojea 
/sig network
/triage accepted
/priority important-soon