@mborsz 2021-12-21
/lgtm
/hold cancel

Thanks, impressive improvement!

Should we open cherry pick requests for older minors?
Such large lock contention (effectively blocking all endpointslice workers) seems to be a bug to me. Also the occurrence of this bug caused by rsyncPeriod (where we add all services to queue, which due to lock contention blocks network programming for O(minutes)) sounds quite severe to me and cherrypick-worth.