@aojea 2021-12-21
> Should we open cherry pick requests for older minors?

I think we should